+++ CHANGELOG.md
@@ -10,6 +10,7 @@ New features:
   character atom type.  The functionality is a significant superset of the previous calculator AI tool which has now been
   removed as a consequence.
 - As we have a new language, added an AIFPL syntax highlighter (this makes the docs render correctly).
+- Removed Claude Sonnet 3.5 support as that is now deprecated by Anthropic.
 
 Bug fixes: